Support longs, shorts, bytes

+++ core/src/main/scala/playground/smithyql/Parser.scala
@@ -97,9 +97,9 @@ object SmithyQLParser {
       (Rfc5234.alpha ~ Parser.charsWhile0(ch => ch.isLetterOrDigit || "_".contains(ch)))
         .map { case (ch, s) => s.prepended(ch) }
 
-    val number: Parser[Int] = Numbers
+    val number: Parser[Long] = Numbers
       .signedIntString
-      .map(_.toInt)
+      .map(_.toLong)
 
     val bool: Parser[Boolean] = string("true").as(true).orElse(string("false").as(false))
